edit botton- ON /styles Css TableAlumnos  modify

+++ app/src/pages/admin/adminAlumnos/editAlumno/editAlumno.jsx
@@ -2,34 +2,30 @@ import React, { useState, useEffect } from "react";
 import "./editAlumno.css"; // Archivo CSS para estilizar el formulario
 import CountryOption from "../../../../components/form/CountryOption";
 import BackLink from "../../../../components/backLink/BackLink";
-import { useNavigate, useLocation} from "react-router-dom";
+import { useNavigate, useLocation } from "react-router-dom";
 import ValidateField from "../../../../components/form/validateField/ValidateField";
-import { deleteAlumno } from "../../../../api/alumnos";
+import { deleteAlumno, updateStudent } from "../../../../api/alumnos";
 
-
-const EditAlumno = ({ onUpdate, onClose }) => {
+const EditAlumno = ({ onUpdate }) => {
   const [formData, setFormData] = useState({
     user_id: "",
     email: "",
     name: "",
     last_name: "",
     nationality: "",
-    cursos: []
   });
 
   const [errors, setErrors] = useState([]);
   const [successMessage, setSuccessMessage] = useState("");
-  const navigate=useNavigate();
+  const navigate = useNavigate();
   const location = useLocation();
-  const {user} = location.state;
+  const { user } = location.state;
 
   const goToStudent = () => {
-     navigate(`/admin/alumnos`);
+    navigate(`/admin/alumnos`);
   };
 
   console.log("User:", user);
-
-
 
   // Pre-rellenar el formulario con los datos actuales del usuario
   useEffect(() => {
@@ -40,7 +36,7 @@ const EditAlumno = ({ onUpdate, onClose }) => {
         name: user.name || "",
         last_name: user.last_name || "",
         nationality: user.nationality || "",
-        cursos: user.cursos || []
+        cursos: user.cursos || [],
       });
     }
   }, [user]);
@@ -54,18 +50,25 @@ const EditAlumno = ({ onUpdate, onClose }) => {
   const handleSubmit = async (e) => {
     e.preventDefault();
     try {
-      await editAlumno(formData.user_id); // Llama a la API para actualizar el usuario
+      await updateStudent(formData); // Pasar los datos actualizados del formulario
       setSuccessMessage("Usuario actualizado con éxito");
       setErrors([]);
       if (onUpdate) onUpdate(); // Llama a la función para refrescar la lista de usuarios
     } catch (error) {
-      setErrors([error.response?.data?.message || "Error al actualizar el usuario"]);
+      setErrors([
+        error.response?.data?.message || "Error al actualizar el usuario",
+      ]);
       setSuccessMessage("");
     }
   };
 
-  const handleDelete = async () =>{
-    try{
+  const handleDelete = async () => {
+    const confirmed = window.confirm(
+      "¿Estás seguro de que deseas borrar este alumno?"
+    );
+    if (!confirmed) return;
+
+    try {
       console.log("Borrando alumno", formData.user_id);
       await deleteAlumno(formData.user_id);
       setFormData({
@@ -74,59 +77,60 @@ const EditAlumno = ({ onUpdate, onClose }) => {
         name: "",
         last_name: "",
         nationality: "",
-        cursos: []
       });
       setSuccessMessage("Alumno eliminado con éxito");
       if (onUpdate) onUpdate();
       goToStudent();
-    }catch (error){
-      setErrors([error.response?.data?.message || "Error al eliminar el alumno"]);
+    } catch (error) {
+      setErrors([
+        error.response?.data?.message || "Error al eliminar el alumno",
+      ]);
       setSuccessMessage("");
     }
-  }
+  };
 
-return (
-  <div>
-    <BackLink title="Volver a Estudiantes" onClick={()=> goToStudent()}/>
-    <div className="edit-user-container">
-      <h2 className="edit-user-title">Editar Usuario</h2>
-      <form onSubmit={handleSubmit} className="edit-user-form">
-        <div className="edit-user-field">
-          <label htmlFor="email">Email:</label>
-          <input
-            id="email"
-            type="email"
-            name="email"
-            value={formData.email}
-            onChange={handleChange}
-            required
-          />
-        </div>
-        <div className="edit-user-field">
-          <label htmlFor="name">Nombre:</label>
-          <input
-            id="name"
-            type="text"
-            name="name"
-            value={formData.name}
-            onChange={handleChange}
-            required
-          />
-        </div>
-        <div className="edit-user-field">
-          <label htmlFor="last_name">Apellido:</label>
-          <input
-            id="last_name"
-            type="text"
-            name="last_name"
-            value={formData.last_name}
-            onChange={handleChange}
-            required
-          />
-        </div>
-        <div className="edit-user-field">
-          <label htmlFor="nationality">Pais de Origen:</label>
-         <select
+  return (
+    <div>
+      <BackLink title="Volver a Estudiantes" onClick={() => goToStudent()} />
+      <div className="edit-user-container">
+        <h2 className="edit-user-title">Editar Usuario</h2>
+        <form onSubmit={handleSubmit} className="edit-user-form">
+          <div className="edit-user-field">
+            <label htmlFor="email">Email:</label>
+            <input
+              id="email"
+              type="email"
+              name="email"
+              value={formData.email}
+              onChange={handleChange}
+              required
+            />
+          </div>
+          <div className="edit-user-field">
+            <label htmlFor="name">Nombre:</label>
+            <input
+              id="name"
+              type="text"
+              name="name"
+              value={formData.name}
+              onChange={handleChange}
+              required
+            />
+          </div>
+          <div className="edit-user-field">
+            <label htmlFor="last_name">Apellido:</label>
+            <input
+              id="last_name"
+              type="text"
+              name="last_name"
+              value={formData.last_name}
+              onChange={handleChange}
+              required
+            />
+          </div>
+          <div className="edit-user-field">
+            <label htmlFor="nationality">Pais de Origen:</label>
+            <select
               type="text"
               name="nationality"
               placeholder="Nacionalidad"
@@ -134,30 +138,36 @@ return (
               onChange={handleChange}
               required
             >
-              <CountryOption value={formData.nationality}/>
+              <CountryOption value={formData.nationality} />
             </select>
-        </div>
-        <button type="submit" className="edit-user-submit" onClick={handleSubmit}>
-          Guardar Cambios
-        </button>
-        <button type="button" className="edit-user-cancel" onClick={handleDelete}>
-          Borrar Alumno
-        </button>
-      </form>
+          </div>
+          <button type="submit" className="edit-user-submit">
+            Guardar Cambios
+          </button>
+          <button
+            type="button"
+            className="edit-user-cancel"
+            onClick={handleDelete}
+          >
+            Borrar Alumno
+          </button>
+        </form>
 
-      {successMessage && <p className="edit-user-success">{successMessage}</p>}
-      {errors.length > 0 && (
-        <div className="edit-user-errors">
-          {errors.map((error, index) => (
-            <p key={index} className="edit-user-error">
-              {error}
-            </p>
-          ))}
-        </div>
-      )}
+        {successMessage && (
+          <p className="edit-user-success">{successMessage}</p>
+        )}
+        {errors.length > 0 && (
+          <div className="edit-user-errors">
+            {errors.map((error, index) => (
+              <p key={index} className="edit-user-error">
+                {error}
+              </p>
+            ))}
+          </div>
+        )}
+      </div>
     </div>
-  </div>
-);
+  );
 };
 
 export default EditAlumno;
